+++ src/ELF/SizingInfo.hpp
@@ -30,6 +30,8 @@ struct sizing_info_t {
   uint64_t gnu_hash = 0;
   uint64_t hash = 0;
   uint64_t rela = 0;
+  uint64_t relr = 0;
+  uint64_t android_rela = 0;
   uint64_t jmprel = 0;
   uint64_t versym = 0;
   uint64_t verdef = 0;